Draft implementation of `--ignoreglyphs` option
As implemented, the option's argument is to be a regex. Any glyphs whose name matches the regex will be ignored for the purposes of fea generation, e.g., omitted from generated classes. if `--ignoreglyphs` is followed immediately by another commandline option, then `^_` is assumed for the regex so all glyphs starting with underscore will be ignored.